+++ ccramic/callbacks/cell_level_callbacks.py
@@ -574,8 +574,12 @@ def init_cell_level_callbacks(dash_app, tmpdirname, authentic_id):
     @dash_app.callback(
         Output("quantification-roi-modal", "is_open"),
         Input('quantify-cur-roi-button', 'n_clicks'),
+        Input('quantify-cur-roi-execute', 'n_clicks'),
         [State("quantification-roi-modal", "is_open")])
-    def toggle_show_quantification_config_modal(n1, is_open):
-        if n1:
-            return not is_open
-        return is_open
+    def toggle_show_quantification_config_modal(n1, execute, is_open):
+        if ctx.triggered_id == "quantify-cur-roi-execute" and execute > 0:
+            return False
+        else:
+            if n1:
+                return not is_open
+            return is_open