+++ tests/test_cell_level_inputs.py
@@ -3,6 +3,9 @@ import os
 import plotly.graph_objs as go
 import pandas as pd
 import plotly
+import pytest
+from dash.exceptions import PreventUpdate
+
 from ccramic.inputs.cell_level_inputs import (
     get_cell_channel_expression_plot,
     generate_umap_plot,
@@ -29,22 +32,36 @@ def test_bar_graph_from_measurements_csv(get_current_dir):
     assert cell_bar_min['layout']['yaxis']['title']['text'] == "min"
     assert cell_bar_min['layout']['title']['text'] == 'Segmented Marker Expression (244 cells)'
 
+    cell_bar_min = get_cell_channel_expression_plot(measurements_csv, mode="min", drop_cols=False)
+    assert cell_bar_min['layout']['xaxis']['title']['text'] == "Channel"
+    assert cell_bar_min['layout']['yaxis']['title']['text'] == "min"
+    assert cell_bar_min['layout']['title']['text'] == 'Segmented Marker Expression (244 cells)'
+
+
 
 def test_bar_graph_from_measurements_csv_with_subsetting(get_current_dir):
     measurements_csv = pd.read_csv(os.path.join(get_current_dir, "cell_measurements.csv"))
     subset_dict = {"x_max": 900, "x_min": 400, "y_max": 65, "y_min": 5}
     cell_bar = get_cell_channel_expression_plot(measurements_csv, subset_dict=subset_dict)
     assert '61 cells' in cell_bar['layout']['title']['text']
+    # assert that no subset is made when the column is not found
+    measurements_csv = measurements_csv.drop(['x_max'], axis=1)
+    cell_bar = get_cell_channel_expression_plot(measurements_csv, subset_dict=subset_dict)
+    assert '244 cells' in cell_bar['layout']['title']['text']
 
 def test_umap_plot(get_current_dir):
     umap_dict = {"UMAP1": [1, 2, 3, 4, 5, 6], "UMAP2": [6, 7, 8, 9, 10, 11]}
     measurements_dict = {"uploads": [os.path.join(get_current_dir, "cell_measurements.csv")]}
     validated_measurements, cols, warning = parse_and_validate_measurements_csv(measurements_dict)
     assert isinstance(warning, dash._callback.NoUpdate)
-    umap_plot = generate_umap_plot(umap_dict, None, validated_measurements, None)
+    umap_plot = generate_umap_plot(umap_dict, "fake_col", validated_measurements, None)
     assert isinstance(umap_plot, plotly.graph_objs._figure.Figure)
     assert umap_plot['layout']['uirevision']
     assert isinstance(generate_umap_plot(None, None, None, None), dash._callback.NoUpdate)
+    umap_plot_2 = generate_umap_plot(umap_dict, "156Gd_FOXA1", validated_measurements, umap_plot)
+    assert isinstance(umap_plot_2, plotly.graph_objs._figure.Figure)
+    assert umap_plot_2['layout']['uirevision']
+
 
 def test_expression_plot_from_interactive_triggers(get_current_dir):
     measurements_dict = {"uploads": [os.path.join(get_current_dir, "cell_measurements.csv")]}
@@ -67,18 +84,36 @@ def test_expression_plot_from_interactive_triggers(get_current_dir):
     interactive_umap, frame = generate_expression_bar_plot_from_interactive_subsetting(validated_measurements, subset_layout,
                                                                                 "mean", subset_layout,
                                                                                 umap_dict, zoom_keys,
-                                                                                "umap-plot")
+                                                                                "umap-plot", category_subset=["test_1"],
+                                                                                category_column="sample",
+                                                                                cols_drop=['sample'])
     assert interactive_umap['layout']['uirevision']
     assert '(0 cells)' in interactive_umap['layout']['title']['text']
+
+    with pytest.raises(PreventUpdate):
+        generate_expression_bar_plot_from_interactive_subsetting(None, subset_layout,
+                                                                 "mean", subset_layout,
+                                                                 umap_dict, zoom_keys,
+                                                                 "umap-plot", category_subset=["test_1"],
+                                                                 category_column="sample",
+                                                                 cols_drop=['sample'])
 
 
 def test_quantification_heatmap(get_current_dir):
     measurements_csv = pd.read_csv(os.path.join(get_current_dir, "cell_measurements.csv"))
-    fig = generate_channel_heatmap(measurements_csv)
+    # assert that the fake column subset will be ignored
+    fig = generate_channel_heatmap(measurements_csv, cols_include=["fake"])
     # assert that the last element in the list of columns in the heatmap is a channel
     assert isinstance(fig, plotly.graph_objs._figure.Figure)
     assert list(fig['data'][0]['x'])[-1] == "209Bi_SMA"
     assert '(244 cells)' in fig['layout']['title']['text']
+    cols_include = ["209Bi_SMA"]
+    fig = generate_channel_heatmap(measurements_csv, cols_include=cols_include)
+    assert '(244 cells)' in fig['layout']['title']['text']
+    # assert that there is only one channel in the entire
+    assert list(fig['data'][0]['x'])[-1] == "209Bi_SMA"
+    assert list(fig['data'][0]['x'])[0] == "209Bi_SMA"
+
 
 def test_heatmap_from_interactive_triggers(get_current_dir):
     measurements_dict = {"uploads": [os.path.join(get_current_dir, "cell_measurements.csv")]}
@@ -104,3 +139,9 @@ def test_heatmap_from_interactive_triggers(get_current_dir):
                                                                         category_column="sample",
                                                                            category_subset=["test_1", "test_2"])
     assert isinstance(interactive_heat, dash._callback.NoUpdate)
+
+    with pytest.raises(PreventUpdate):
+        generate_heatmap_from_interactive_subsetting(None, subset_layout,
+                                                     embeddings, zoom_keys, "umap-projection-options",
+                                                     category_column="sample",
+                                                     category_subset=["test_1", "test_2"])